style(header): sort imports and format new useSearchBar hook
