+++ src/components/layout/Header/Header.tsx
@@ -1,8 +1,5 @@
-import { useQueryClient } from '@tanstack/react-query'
 import { LogIn, LogOut, Menu, Search, User } from 'lucide-react'
-import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
-import { useNavigate } from 'react-router-dom'
 
 import { Button } from '@/components/ui/button'
 import {
@@ -14,9 +11,8 @@ import {
 import { Input } from '@/components/ui/input'
 import { LanguageSelector } from '@/components/ui/LanguageSelector'
 import { SpotifyIcon } from '@/components/ui/SpotifyIcon'
-import { useDebounce } from '@/hooks/useDebounce'
+import { useSearchBar } from '@/hooks/useSearchBar'
 import { useSpotifyAuth } from '@/hooks/useSpotifyAuth'
-import { useSearchStore } from '@/stores/searchStore'
 
 export interface HeaderProps {
   onMenuToggle?: () => void
@@ -25,52 +21,9 @@ export interface HeaderProps {
 
 export function Header({ onMenuToggle, searchPlaceholder }: HeaderProps) {
   const { t } = useTranslation()
-  const navigate = useNavigate()
-  const queryClient = useQueryClient()
   const { isAuthenticated, login, logout } = useSpotifyAuth()
-  const { searchQuery, setSearchQuery, setDebouncedSearchQuery } =
-    useSearchStore()
-
-  // Debounce the search query with 350ms delay
-  const debouncedSearchQuery = useDebounce(searchQuery, 350)
-
-  // Update debounced search query in store when it changes
-  useEffect(() => {
-    setDebouncedSearchQuery(debouncedSearchQuery)
-  }, [debouncedSearchQuery, setDebouncedSearchQuery])
-
-  const handleSearchChange = (value: string) => {
-    setSearchQuery(value)
-
-    // If input is cleared, invalidate queries
-    if (!value.trim()) {
-      queryClient.removeQueries({
-        queryKey: ['spotify-search'],
-      })
-      queryClient.removeQueries({
-        queryKey: ['spotify-search-by-type'],
-      })
-    }
-
-    // Navigate to search page if user types something
-    if (value.trim()) {
-      const queryParams = new URLSearchParams({
-        q: value,
-        market: 'BR',
-      })
-      navigate(`/search?${queryParams.toString()}`)
-    }
-  }
-
-  const handleSearchKeyPress = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter' && searchQuery.trim()) {
-      const queryParams = new URLSearchParams({
-        q: searchQuery,
-        market: 'BR',
-      })
-      navigate(`/search?${queryParams.toString()}`)
-    }
-  }
+  const { searchQuery, handleSearchChange, handleSearchKeyPress } =
+    useSearchBar()
 
   return (
     <header
